move middleware out to it's own file. conditionally use middleware based on env
