+++ packages/gatsby-tinacms-auth/contants.ts
@@ -1,2 +1,3 @@
 export const VIRTUAL_SERVICE_DOMAIN = 'tina.io'
 export const TINA_CONNECTOR_ID = 'eb1c069d-8467-4029-927c-53a95b0de514'
+export const AUTH_COOKIE_KEY = 'tina-auth'
